bug: forgot password fixed
