+++ src/App.js
@@ -6,11 +6,12 @@ import SigninPage from './pages/signin';
 import Error404 from './pages/Error404';
 import BlogPage from './pages/blogs';
 import FeedbackPage from './pages/feedback';
-import Forgot_password from './pages/ForgotPassword';
 import SignupPage from './pages/signup';
 import Profile from './pages/profile';
 import Navbar from './Components/Navbar';
 import ContactPage from './pages/contact';
+import GetMail from './Components/Forgot_password/ForgotPassword';
+import Verify from './Components/Forgot_password/OTPVerfication';
 
 // import Contact from "./Components/Contactus/Contact";
 // window.$crisp=[];
@@ -30,7 +31,12 @@ function App() {
 				<Route
 					exact
 					path="/signin/forgotPassword"
-					element={<Forgot_password />}
+					element={<GetMail/>}
+				/>
+				<Route
+					exact
+					path="/signin/otpverification"
+					element={<Verify/>}
 				/>
 				<Route path="/*" element={<Error404 />} />
 				<Route exact path="/signup" element={<SignupPage />} />
