+++ src/Components/Forgot_password/Forgot_password.css
@@ -41,3 +41,7 @@
 .forgot_resend:hover {
   color: white;
 }
+
+.email-form{
+  margin-top: 10%;
+}
