+++ src/pages/ForgotPassword.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Forgot_password from '../Components/Forgot_password/Forgot_password';
+import Forgot_password from '../Components/Forgot_password/OTPVerfication';
 
 const SigninPage = () => {
   return (
